BexDrive startService/endservice/deliverpackages update

+++ BexDrive/app/src/main/java/com/example/bexdrive/entity/Location.kt
@@ -5,6 +5,6 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 class Location (
-    val Latitude : Int,
-    val Longitude : Int
+    val Latitude : Double,
+    val Longitude : Double
 ) : Parcelable
